requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.5
flask-wtf==1.1.1
python-dotenv==1.0.1
twilio==8.12.0
openai==1.12.0
google-generativeai==0.3.0
schedule==1.2.1
werkzeug==2.3.8
requests==2.31.0
gunicorn
